+++ packages/main.ts
@@ -41,10 +41,10 @@ const config: IImperativeConfig = {
         Imperative.parse();
         timingApi.mark("AFTER_PARSE");
         timingApi.measure("Imperative.parse", "BEFORE_PARSE", "AFTER_PARSE");
-    }
-    catch (initErr) {
+    } catch (initErr) {
         Imperative.console.fatal("Error initializing " + Constants.DISPLAY_NAME +
             ":\n "
             + inspect(initErr));
+        process.exit(1);
     }
 })();