+++ packages/zosfiles/__tests__/__system__/cli/hRecall/ds/cli.files.HRecall.ds.system.test.ts
@@ -18,6 +18,7 @@ import { Session, Imperative } from "@zowe/imperative";
 import { List, Delete, Create, CreateDataSetTypeEnum, IListOptions } from "../../../../..";
 import { IRecallOptions } from "../../../../../src/api/methods/hRecall/doc/IRecallOptions";
 import { inspect } from "util";
+import { HMigrate } from "../../../../../src/api";
 
 let TEST_ENVIRONMENT: ITestEnvironment;
 let defaultSystem: ITestPropertiesSchema;
@@ -69,6 +70,7 @@ describe("Recall Dataset", () => {
       beforeEach(async () => {
         try {
           await Create.dataSet(REAL_SESSION, CreateDataSetTypeEnum.DATA_SET_SEQUENTIAL, dataSetName1);
+          await HMigrate.dataSet(REAL_SESSION, dataSetName1);
         } catch (err) {
           Imperative.console.info(`Error: ${inspect(err)}`);
         }
@@ -109,6 +111,7 @@ describe("Recall Dataset", () => {
       beforeEach(async () => {
         try {
           await Create.dataSet(REAL_SESSION, CreateDataSetTypeEnum.DATA_SET_PARTITIONED, dataSetName2);
+          await HMigrate.dataSet(REAL_SESSION, dataSetName3);
         } catch (err) {
           Imperative.console.info(`Error: ${inspect(err)}`);
         }
@@ -151,12 +154,13 @@ describe("Recall Dataset", () => {
       beforeEach(async () => {
         try {
           await Create.dataSet(REAL_SESSION, CreateDataSetTypeEnum.DATA_SET_SEQUENTIAL, dataSetName3);
+          await HMigrate.dataSet(REAL_SESSION, dataSetName3);
         } catch (err) {
           Imperative.console.info(`Error: ${inspect(err)}`);
         }
       });
       it("Should throw an error if a missing data set name is selected", async () => {
-        const response = runCliScript(recallScript, TEST_ENVIRONMENT, ["MISSING.DATA.SET", dataSetName3]);
+        const response = runCliScript(recallScript, TEST_ENVIRONMENT, ["", dataSetName3]);
 
         expect(response.stderr.toString()).toBeTruthy();
         expect(response.status).toBe(1);