+++ packages/zosfiles/src/api/constants/ZosFiles.messages.ts
@@ -48,6 +48,14 @@ export const ZosFilesMessages: { [key: string]: IMessageDefinition } = {
     },
 
     /**
+     * Message indicating that the data set "like" name is required
+     * @type {IMessageDefinition}
+     */
+    missingDatasetLikeName: {
+        message: "Specify the name of the data set to \"allocate like\" from."
+    },
+
+    /**
      * Message indicating that the USS File name is required
      * @type {IMessageDefinition}
      */
@@ -524,14 +532,6 @@ export const ZosFilesMessages: { [key: string]: IMessageDefinition } = {
     },
 
     /**
-     * Message indicating that the data set was recalled successfully.
-     * @type {IMessageDefinition}
-     */
-    datasetRecallRequested: {
-        message: "Data set recall requested."
-    },
-
-    /**
      * Message indicating that the data set has been renamed successfully.
      * @type {IMessageDefinition}
      */
@@ -552,7 +552,15 @@ export const ZosFilesMessages: { [key: string]: IMessageDefinition } = {
      * @type {IMessageDefinition}
      */
     datasetMigrationRequested: {
-        message: "Data set migraton requested."
+        message: "Data set migration requested."
+    },
+
+    /**
+     * Message indicating that the data set was recalled successfully.
+     * @type {IMessageDefinition}
+     */
+    datasetRecallRequested: {
+        message: "Data set recall requested."
     },
 
     /**